+++ changelog.md
@@ -4,7 +4,7 @@ _(Note: This changelog focuses on the major changes between the different
 versions. Therefore, it may not contain all changes. Especially smaller fixes or
 improvements may be omitted.)_
 
-## Next version (2020-09-??)
+## Version 0.1.10 (2020-08-30)
 
 * COVID-19 data will now show the numbers of the last ten days instead of just
   the last seven days.